requirements.txt
appdirs==1.4.3
awsebcli==3.10.1
backports.ssl-match-hostname==3.5.0.1
blessed==1.14.2
boto==2.28.0
botocore==1.5.44
cement==2.8.2
colorama==0.3.7
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.2
MarkupSafe==0.23
p2p-latimes==2.0.30
packaging==16.8
pathspec==0.5.0
pkg-resources==0.0.0
PyMySQL==0.6.3
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
SQLAlchemy==0.9.8
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.40.0
Werkzeug==0.9.6
WTForms==2.0.1
